2018.03.08 19:37

+++ ByTime/2019/02/25/sanrd.cpp
@@ -9,11 +9,11 @@
 using std::swap;using std::max;using std::copy;
 using std::cerr;using std::min;using std::fill;
 
-const int MAXN = 10009;
+const short MAXN = 10009;
 
-int f[MAXN][MAXN];
-int s[MAXN];
-int n;
+short f[MAXN][MAXN];
+short s[MAXN];
+short n;
 
 int main()
 {
@@ -24,12 +24,12 @@ int main()
 		while (s[0]<1 || s[0]>9);
 	do {s[++n]=getchar()-'0';}
 		while (s[n]>=1 && s[n]<=9);
-	for (int i=0;i<=n;i++)
+	for (short i=0;i<=n;i++)
 		f[i][i]=0;
-	for (int k=1;k<=n;k++)
-		for (int i=0;i<=n && i+k<=n;i++)
-			for (int p=i;p<i+k;p++)
-				f[i][i+k]=min(f[i][i+k],max(f[i][p],f[p+1][i+k])+s[p]);
+	for (short k=1;k<=n;k++)
+		for (short i=0;i<=n && i+k<=n;i++)
+			for (short p=i;p<i+k;p++)
+				f[i][i+k]=min(f[i][i+k],(short)(max(f[i][p],f[p+1][i+k])+s[p]));
 	printf("%d\n",f[0][n]);
 	return 0;
 }
